requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
langdetect
